+++ wger_api_client/api/user_trophy/user_trophy_list.py
@@ -64,9 +64,9 @@ def _get_kwargs(
 
     params["id"] = id
 
-    json_id_in: list[int] | Unset = UNSET
+    json_id_in: str | Unset = UNSET
     if not isinstance(id_in, Unset):
-        json_id_in = id_in
+        json_id_in = ",".join(str(v) for v in id_in)
 
     params["id__in"] = json_id_in
 
@@ -80,9 +80,9 @@ def _get_kwargs(
 
     params["trophy"] = trophy
 
-    json_trophy_in: list[int] | Unset = UNSET
+    json_trophy_in: str | Unset = UNSET
     if not isinstance(trophy_in, Unset):
-        json_trophy_in = trophy_in
+        json_trophy_in = ",".join(str(v) for v in trophy_in)
 
     params["trophy__in"] = json_trophy_in
 
@@ -98,9 +98,9 @@ def _get_kwargs(
 
     params["trophy__trophy_type"] = json_trophy_trophy_type
 
-    json_trophy_trophy_type_in: list[str] | Unset = UNSET
+    json_trophy_trophy_type_in: str | Unset = UNSET
     if not isinstance(trophy_trophy_type_in, Unset):
-        json_trophy_trophy_type_in = trophy_trophy_type_in
+        json_trophy_trophy_type_in = ",".join(str(v) for v in trophy_trophy_type_in)
 
     params["trophy__trophy_type__in"] = json_trophy_trophy_type_in